feat(frontend): rename files ss-505
