+++ apps/frontend/src/pages/manage-categories/libs/components/create-category-modal/libs/constants/constants.ts
@@ -1,1 +1,1 @@
-export { DEFAULT_CREATE_CATEGORY_PAYLOAD } from "./default-create-category-payload.js";
+export { DEFAULT_CREATE_CATEGORY_PAYLOAD } from "./default-create-category-payload.constant.js";
