+++ apps/frontend/src/pages/manage-categories/libs/components/create-category-modal/libs/constants/default-create-category-payload.constant.ts
