+++ apps/frontend/src/pages/manage-categories/libs/components/create-category-modal/libs/types/create-category-form-values.type.ts
