+++ apps/frontend/src/pages/manage-categories/libs/components/create-category-modal/libs/types/types.ts
@@ -1,1 +1,1 @@
-export { type CreateCategoryFormValues } from "./create-category-form-values.js";
+export { type CreateCategoryFormValues } from "./create-category-form-values.type.js";
